Fix markdown table emoji paint coverage
