+++ src/agent.ts
@@ -117,6 +117,8 @@ export type {
   TuiMarkdownBlock,
   TuiMarkdownInlineSegment,
   TuiMarkdownNode,
+  TuiMarkdownTableCell,
+  TuiMarkdownTableCellAlign,
   TuiMarkdownVisualRow,
   TuiMarkdownVisualSegment,
 } from "./vue/markdown/types.js";
