+++ src/core/terminal/create-terminal.ts
@@ -1,5 +1,6 @@
 import type { GridBuffer } from "../buffer/buffer.js";
 import type { TerminalRenderPlane, TerminalRenderPlanes } from "../render-plane.js";
+import { segmentedGraphemes } from "../../utils/grapheme.js";
 import type {
   BufferSnapshot,
   Cell,
@@ -30,15 +31,6 @@ import {
 } from "../buffer/buffer.js";
 import { TERMINAL_RENDER_PLANES } from "../render-plane.js";
 import { Emitter } from "./emitter.js";
-
-type GraphemeSegment = Readonly<{ segment: string }>;
-type GraphemeSegmenter = Readonly<{ segment(input: string): Iterable<GraphemeSegment> }>;
-type IntlWithSegmenter = typeof Intl & {
-  Segmenter?: new (
-    locales?: string | string[],
-    options?: Readonly<{ granularity?: "grapheme" }>,
-  ) => GraphemeSegmenter;
-};
 
 function isControlChar(ch: string): boolean {
   return ch === "\n" || ch === "\r" || ch === "\t";
@@ -272,114 +264,6 @@ function normalizeDirtyRows(
     out.push(yy);
   }
   return out;
-}
-
-// Shared Intl.Segmenter instance for grapheme cluster iteration
-let sharedGraphemeSegmenter: GraphemeSegmenter | null = null;
-try {
-  const Segmenter = (Intl as IntlWithSegmenter).Segmenter;
-  sharedGraphemeSegmenter = Segmenter
-    ? new Segmenter(undefined, { granularity: "grapheme" })
-    : null;
-} catch {
-  // Fall back to code point iteration if Intl.Segmenter is not available
-}
-
-function needsGraphemeSegmentation(text: string): boolean {
-  for (const ch of text) {
-    const cp = ch.codePointAt(0)!;
-    if (cp === 0x200d) return true;
-    if ((cp >= 0xfe00 && cp <= 0xfe0f) || (cp >= 0xe0100 && cp <= 0xe01ef)) return true;
-    if (
-      (cp >= 0x0300 && cp <= 0x036f) ||
-      (cp >= 0x1ab0 && cp <= 0x1aff) ||
-      (cp >= 0x1dc0 && cp <= 0x1dff) ||
-      (cp >= 0x20d0 && cp <= 0x20ff) ||
-      (cp >= 0xfe20 && cp <= 0xfe2f)
-    ) {
-      return true;
-    }
-    if (cp >= 0x1f3fb && cp <= 0x1f3ff) return true;
-    if (cp >= 0x1f1e6 && cp <= 0x1f1ff) return true;
-    if (cp >= 0xe0000 && cp <= 0xe007f) return true;
-  }
-  return false;
-}
-
-function isVariationSelector(codePoint: number): boolean {
-  return (
-    (codePoint >= 0xfe00 && codePoint <= 0xfe0f) || (codePoint >= 0xe0100 && codePoint <= 0xe01ef)
-  );
-}
-
-function isCombiningMark(codePoint: number): boolean {
-  return (
-    (codePoint >= 0x0300 && codePoint <= 0x036f) ||
-    (codePoint >= 0x1ab0 && codePoint <= 0x1aff) ||
-    (codePoint >= 0x1dc0 && codePoint <= 0x1dff) ||
-    (codePoint >= 0x20d0 && codePoint <= 0x20ff) ||
-    (codePoint >= 0xfe20 && codePoint <= 0xfe2f)
-  );
-}
-
-function isEmojiModifier(codePoint: number): boolean {
-  return codePoint >= 0x1f3fb && codePoint <= 0x1f3ff;
-}
-
-function isRegionalIndicator(codePoint: number): boolean {
-  return codePoint >= 0x1f1e6 && codePoint <= 0x1f1ff;
-}
-
-function isEmojiTag(codePoint: number): boolean {
-  return codePoint >= 0xe0000 && codePoint <= 0xe007f;
-}
-
-function fallbackGraphemeSegments(text: string): readonly GraphemeSegment[] {
-  const out: GraphemeSegment[] = [];
-  let segment = "";
-  let regionalIndicators = 0;
-  let joinNext = false;
-
-  const pushSegment = () => {
-    if (segment) out.push({ segment });
-  };
-
-  for (const ch of text) {
-    const codePoint = ch.codePointAt(0)!;
-    const attach =
-      segment !== "" &&
-      (joinNext ||
-        codePoint === 0x200d ||
-        isVariationSelector(codePoint) ||
-        isCombiningMark(codePoint) ||
-        isEmojiModifier(codePoint) ||
-        isEmojiTag(codePoint) ||
-        (isRegionalIndicator(codePoint) && regionalIndicators === 1));
-
-    if (!attach) {
-      pushSegment();
-      segment = ch;
-      regionalIndicators = isRegionalIndicator(codePoint) ? 1 : 0;
-    } else {
-      segment += ch;
-      if (isRegionalIndicator(codePoint)) regionalIndicators++;
-      else if (!isVariationSelector(codePoint) && !isCombiningMark(codePoint))
-        regionalIndicators = 0;
-    }
-
-    if (codePoint === 0x200d) joinNext = true;
-    else if (joinNext && !isVariationSelector(codePoint) && !isCombiningMark(codePoint)) {
-      joinNext = false;
-    }
-  }
-
-  pushSegment();
-  return out;
-}
-
-function segmentedGraphemes(text: string): Iterable<GraphemeSegment> | null {
-  if (!needsGraphemeSegmentation(text)) return null;
-  return sharedGraphemeSegmenter?.segment(text) ?? fallbackGraphemeSegments(text);
 }
 
 export function getPlaneTerminal(terminal: Terminal, plane: TerminalRenderPlane): Terminal {
