+++ src/vue/utils/text.ts
@@ -1,19 +1,5 @@
 import { charCellWidth, type WidthProvider } from "../../core/buffer/width.js";
-
-interface GraphemeSegment {
-  segment: string;
-  index: number;
-}
-
-type GraphemeSegmenter = {
-  segment(input: string): Iterable<GraphemeSegment>;
-};
-type IntlWithSegmenter = typeof Intl & {
-  Segmenter?: new (
-    locales?: string | string[],
-    options?: Readonly<{ granularity?: "grapheme" }>,
-  ) => GraphemeSegmenter;
-};
+import { segmentedGraphemes } from "../../utils/grapheme.js";
 
 export interface TextCellSegment {
   text: string;
@@ -69,129 +55,6 @@ function isAscii(text: string): boolean {
     if (text.charCodeAt(i) > 0x7f) return false;
   }
   return true;
-}
-
-function needsGraphemeSegmentation(text: string): boolean {
-  // Be conservative: return true if we see characters that commonly participate in multi-codepoint
-  // grapheme clusters (combining marks, ZWJ sequences, emoji variation selectors, skin tone modifiers,
-  // regional-indicator flags, or emoji tag sequences). If none are present, code-point iteration is
-  // generally safe.
-  for (const ch of text) {
-    const cp = ch.codePointAt(0)!;
-    // ZWJ
-    if (cp === 0x200d) return true;
-    // Variation selectors (incl. VS16)
-    if ((cp >= 0xfe00 && cp <= 0xfe0f) || (cp >= 0xe0100 && cp <= 0xe01ef)) return true;
-    // Combining diacritics + common marks blocks
-    if (
-      (cp >= 0x0300 && cp <= 0x036f) ||
-      (cp >= 0x1ab0 && cp <= 0x1aff) ||
-      (cp >= 0x1dc0 && cp <= 0x1dff) ||
-      (cp >= 0x20d0 && cp <= 0x20ff) ||
-      (cp >= 0xfe20 && cp <= 0xfe2f)
-    ) {
-      return true;
-    }
-    // Emoji modifiers (skin tones)
-    if (cp >= 0x1f3fb && cp <= 0x1f3ff) return true;
-    // Regional indicator symbols (flags are pairs)
-    if (cp >= 0x1f1e6 && cp <= 0x1f1ff) return true;
-    // Emoji tag sequences, including subdivision flags
-    if (cp >= 0xe0000 && cp <= 0xe007f) return true;
-  }
-  return false;
-}
-
-function isVariationSelector(codePoint: number): boolean {
-  return (
-    (codePoint >= 0xfe00 && codePoint <= 0xfe0f) || (codePoint >= 0xe0100 && codePoint <= 0xe01ef)
-  );
-}
-
-function isCombiningMark(codePoint: number): boolean {
-  return (
-    (codePoint >= 0x0300 && codePoint <= 0x036f) ||
-    (codePoint >= 0x1ab0 && codePoint <= 0x1aff) ||
-    (codePoint >= 0x1dc0 && codePoint <= 0x1dff) ||
-    (codePoint >= 0x20d0 && codePoint <= 0x20ff) ||
-    (codePoint >= 0xfe20 && codePoint <= 0xfe2f)
-  );
-}
-
-function isEmojiModifier(codePoint: number): boolean {
-  return codePoint >= 0x1f3fb && codePoint <= 0x1f3ff;
-}
-
-function isRegionalIndicator(codePoint: number): boolean {
-  return codePoint >= 0x1f1e6 && codePoint <= 0x1f1ff;
-}
-
-function isEmojiTag(codePoint: number): boolean {
-  return codePoint >= 0xe0000 && codePoint <= 0xe007f;
-}
-
-function fallbackGraphemeSegments(text: string): readonly GraphemeSegment[] {
-  const out: GraphemeSegment[] = [];
-  let segment = "";
-  let segmentStart = 0;
-  let pos = 0;
-  let regionalIndicators = 0;
-  let joinNext = false;
-
-  const pushSegment = () => {
-    if (segment) out.push({ segment, index: segmentStart });
-  };
-
-  for (const ch of text) {
-    const codePoint = ch.codePointAt(0)!;
-    const chStart = pos;
-    pos += ch.length;
-
-    const attach =
-      segment !== "" &&
-      (joinNext ||
-        codePoint === 0x200d ||
-        isVariationSelector(codePoint) ||
-        isCombiningMark(codePoint) ||
-        isEmojiModifier(codePoint) ||
-        isEmojiTag(codePoint) ||
-        (isRegionalIndicator(codePoint) && regionalIndicators === 1));
-
-    if (!attach) {
-      pushSegment();
-      segment = ch;
-      segmentStart = chStart;
-      regionalIndicators = isRegionalIndicator(codePoint) ? 1 : 0;
-    } else {
-      segment += ch;
-      if (isRegionalIndicator(codePoint)) regionalIndicators++;
-      else if (!isVariationSelector(codePoint) && !isCombiningMark(codePoint))
-        regionalIndicators = 0;
-    }
-
-    if (codePoint === 0x200d) joinNext = true;
-    else if (joinNext && !isVariationSelector(codePoint) && !isCombiningMark(codePoint)) {
-      joinNext = false;
-    }
-  }
-
-  pushSegment();
-  return out;
-}
-
-function segmentedGraphemes(text: string): Iterable<GraphemeSegment> | null {
-  if (!needsGraphemeSegmentation(text)) return null;
-  if (graphemeSegmenter) return graphemeSegmenter.segment(text) as any as Iterable<GraphemeSegment>;
-  return fallbackGraphemeSegments(text);
-}
-
-let graphemeSegmenter: GraphemeSegmenter | null = null;
-try {
-  // Some runtimes may not support Intl.Segmenter; fall back to code-point iteration.
-  const Segmenter = typeof Intl !== "undefined" ? (Intl as IntlWithSegmenter).Segmenter : undefined;
-  graphemeSegmenter = Segmenter ? new Segmenter(undefined, { granularity: "grapheme" }) : null;
-} catch {
-  graphemeSegmenter = null;
 }
 
 function forEachGrapheme(text: string, cb: (g: string) => void | false): void {
